requirements.txt
openai
requests
python-dotenv
streamlit
pydantic